requirements.txt
asgiref==3.7.2
attrs==20.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
django-environ==0.11.2
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
jellyfish==0.11.2
lib50==3.0.7
markdown2==2.4.10
outcome==1.2.0
pexpect==4.8.0
Pillow==10.1.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py3dns==4.0.0
pycparser==2.21
PySocks==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
submit50==3.1.1
termcolor==1.1.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
urllib3==2.0.4
validate-email==1.3
wsproto==1.2.0
